requirements.txt
psycopg2-binary==2.8.5
requests==2.23.0
pytz==2020.1

